pyproject.toml
[project]
name = "aigtd"
version = "0.1.0"
description = "AI Goal Tracking and Development"
requires-python = ">=3.8"
dependencies = [
    "Flask==2.3.3",
    "Flask-CORS==4.0.0",
    "SQLAlchemy==2.0.23",
    "Flask-SQLAlchemy==3.1.1",
    "python-dotenv==1.0.0",
    "openai>=1.10.0",
    "requests==2.31.0",
    "langchain-openai==0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []
